@@ -14,28 +14,38 @@
     "vccw:remove": "rm -rf ./vccw",
     "vccw": "yarn run vccw:clone && yarn run vccw:checkout && yarn run vccw:copy && yarn run vccw:remove",
     "setup": "yarn run build && yarn run vccw && vagrant up && yarn run start",
-    "create": "node tools/create/createMessage.js && yarn install && node tools/create/create.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "create": "node ./tools/create/createMessage.js && yarn install && node ./tools/create/create.js",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "analyze-bundle": "babel-node ./tools/analyzeBundle.js"
   },
   "dependencies": {
-    "babel-core": "^6.0.20",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-es2015": "^6.0.15",
-    "babel-preset-react": "^6.0.15",
-    "babel-preset-stage-0": "^6.0.15",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-ga": "^2.1.2",
+    "moment": "^2.18.1",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "react-ga": "^2.2.0",
     "react-redux": "^5.0.1",
     "react-router": "^3.0.0",
     "react-router-redux": "^4.0.7",
     "redux": "^3.6.0",
     "redux-thunk": "^2.1.0",
+    "reset-css": "^2.2.0",
     "wpapi": "^1.0.0"
   },
   "devDependencies": {
-    "autoprefixer": "^7.1.2",
-    "babel-loader": "^6.0.1",
+    "autoprefixer": "7.1.2",
+    "babel-cli": "6.24.1",
+    "babel-core": "6.25.0",
+    "babel-loader": "7.1.1",
+    "babel-plugin-react-css-modules": "^3.1.0",
+    "babel-plugin-transform-react-constant-elements": "6.23.0",
+    "babel-plugin-transform-react-jsx": "^6.24.1",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.6",
+    "babel-polyfill": "6.23.0",
+    "babel-preset-env": "1.6.0",
+    "babel-preset-es2015": "6.24.1",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-0": "6.24.1",
     "browser-sync": "^2.18.2",
     "bs-html-injector": "^3.0.3",
     "chalk": "^2.0.1",
@@ -43,27 +53,32 @@
     "copy": "^0.3.0",
     "css-loader": "^0.28.4",
     "del": "^2.2.2",
+    "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^0.11.2",
-    "fs": "^0.0.1-security",
     "image-webpack-loader": "^3.3.1",
     "json-loader": "^0.5.4",
     "node-sass": "^4.5.3",
     "postcss-loader": "^2.0.6",
+    "postcss-nested": "^2.0.2",
+    "postcss-scss": "^1.0.2",
+    "postcss-unprefix": "^2.1.1",
     "prompt": "^1.0.0",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "redux-devtools": "^3.3.1",
-    "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-devtools-log-monitor": "^1.1.1",
+    "react-hot-loader": "^3.0.0-beta.7",
+    "redux-devtools": "^3.4.0",
+    "redux-devtools-dock-monitor": "^1.1.2",
+    "redux-devtools-log-monitor": "^1.3.0",
     "replace": "^0.3.0",
+    "rimraf": "^2.6.1",
     "sass-loader": "^6.0.6",
     "standard": "^8.3.0",
     "standard-loader": "^5.0.0",
     "style-loader": "^0.18.2",
     "url-loader": "^0.5.9",
-    "webpack": "^1.13.3",
-    "webpack-dev-middleware": "^1.8.4",
-    "webpack-hot-middleware": "^2.13.2",
-    "write-file-webpack-plugin": "^3.4.2"
+    "webpack": "3.2.0",
+    "webpack-bundle-analyzer": "2.8.2",
+    "webpack-dev-middleware": "1.11.0",
+    "webpack-hot-middleware": "2.18.2",
+    "webpack-md5-hash": "0.0.5"
   },
   "keywords": [
     "react",
